Bugfixe und Erweiterungen Fähigkeiten bearbeiten + Ausloggen

+++ frontend/src/app/modules/profile/components/kex-profile/kex-profile-skills/kex-profile-skill/kex-profile-skill.component.ts
@@ -1,10 +1,24 @@
 import {Component, EventEmitter, Input, OnDestroy, OnInit, Output} from '@angular/core';
-import {KexUserSkill} from "../../../../models/kex-profile.model";
+import {KexSkill, KexUserSkill} from "../../../../models/kex-profile.model";
 import {KexCoreService} from "../../../../../../core/services/kex-core.service";
 import {KexProfileService} from "../../../../services/kex-profile.service";
-import {Observable, Subscription} from "rxjs";
+import {debounceTime, distinctUntilChanged, filter, Observable, of, Subscription, switchMap} from "rxjs";
 import {state} from "@angular/animations";
 import {KexLoadState} from "../../../../../../core/models/kex-core.models";
+import {FormControl} from "@angular/forms";
+import {KexProfileConnectorService} from "../../../../services/kex-profile-connector.service";
+import {Store} from "@ngrx/store";
+import {KexProfileState} from "../../../../store/kex-profile.state";
+import {
+  AddSkillActions,
+  DeleteSkillActions,
+  EditSkillActions,
+  UpdateVisibilitySkillActions
+} from "../../../../store/actions/kex-profile.actions";
+import {MatDialog} from "@angular/material/dialog";
+import {
+  KexModalConfirmationComponent
+} from "../../../../../../shared/components/kex-modal/kex-modal-confirmation/kex-modal-confirmation.component";
 
 @Component({
   selector: 'kex-profile-skill',
@@ -15,19 +29,39 @@ export class KexProfileSkillComponent implements OnInit, OnDestroy {
   @Input() userSkill: KexUserSkill | undefined;
   @Output() leaveNewSkillMode = new EventEmitter<boolean>();
 
-  title = '';
+  private subscriptions: Subscription[] = [];
   level = 1;
   visible = false;
-
   editMode = false;
-  private subscriptions: Subscription[] = [];
+  controlTitle = new FormControl('');
+  suggestionSkillList: KexSkill[] = [];
+  addNewSkill = false;
 
   constructor(private coreService: KexCoreService,
-              private profileService: KexProfileService) {
+              private profileService: KexProfileService,
+              private profileConnectorService: KexProfileConnectorService,
+              private store : Store<KexProfileState>,
+              public dialog: MatDialog) {
   }
 
   get color() {
     return this.visible ? 'primary' : 'accent';
+  }
+
+  get title(): string {
+    return this.userSkill?.skill.title || '';
+  }
+
+  observeSuggestionForTitle() {
+    this.controlTitle.valueChanges.pipe(
+      filter(data => data != null && data.trim().length > 2),
+      debounceTime(100),
+      switchMap((value) => {
+        return value ? this.profileConnectorService.getSkillSuggestions(value) : of([]);
+      })
+    ).subscribe(data => {
+      this.suggestionSkillList = data;
+    })
   }
 
   ratingUpdated(rating: number) {
@@ -39,13 +73,22 @@ export class KexProfileSkillComponent implements OnInit, OnDestroy {
   }
 
   deleteSkill() {
-    if (this.userSkill) {
-      this.profileService.deleteSkill(this.userSkill);
-    }
+
+        const dialogRef = this.dialog.open(KexModalConfirmationComponent, {
+          data: {labelAction: 'Löschen',
+            labelHeadline: 'Fähigkeit löschen',
+            labelDescription: 'Soll die Fähigkeit wirklich gelöscht werden? Die Aktion kann nicht wieder rückgängig gemacht werden.'},
+        });
+
+        dialogRef.afterClosed().subscribe(result => {
+          if(result && this.userSkill){
+            this.profileService.deleteSkill(this.userSkill);
+          }
+        });
   }
 
   leaveEditMode() {
-    if (this.userSkill) {
+    if (!this.addNewSkill) {
       this.editMode = false;
     } else {
       this.leaveNewSkillMode.emit(true);
@@ -61,32 +104,45 @@ export class KexProfileSkillComponent implements OnInit, OnDestroy {
   }
 
   saveSkill(): void {
-    if (this.userSkill) {
-      const skill: KexUserSkill = {
-        ...this.userSkill,
-        skill: {id: 1, title: this.title},
-        level: this.level,
-        visible: this.visible
-      };
-      this.profileService.saveSkill(skill);
-    } else {
-      const skill: KexUserSkill = {id: 0, skill: {id: 1, title: this.title}, level: this.level, visible: this.visible};
-      this.profileService.addSkill(skill);
+    let titleNew: string | null = this.controlTitle.value;
+    if (titleNew != null) {
+      let skillId: number = this.userSkill && this.userSkill.skill.title == titleNew ? this.userSkill.skill.id || 0 : 0;
+      if (this.userSkill) {
+        const skill: KexUserSkill = {
+          ...this.userSkill,
+          skill: {id: skillId, title: titleNew},
+          level: this.level,
+          visible: this.visible
+        };
+        this.userSkill = skill;
+        this.profileService.saveSkill(skill);
+      } else {
+        const skill: KexUserSkill = {
+          id: 0,
+          skill: {id: skillId, title: titleNew},
+          level: this.level,
+          visible: this.visible
+        };
+        this.userSkill = skill;
+        this.profileService.addSkill(skill);
+      }
     }
   }
 
   ngOnInit(): void {
     if (this.userSkill) {
-      this.title = this.userSkill.skill.title;
+      this.controlTitle.patchValue(this.userSkill.skill.title);
       this.level = this.userSkill.level;
       this.visible = this.userSkill.visible;
     } else {
+      this.addNewSkill = true;
       this.editMode = true;
     }
     this.observeEditSkill();
     this.observeDeleteSkill();
     this.observeAddSkill();
     this.observeUpdateVisibilitySkill();
+    this.observeSuggestionForTitle();
   }
 
   get $deleteSkillLoadState(): Observable<KexLoadState> {
@@ -99,7 +155,9 @@ export class KexProfileSkillComponent implements OnInit, OnDestroy {
       ).subscribe(state => this.coreService.handleRequestState(state,
           'Fähigkeit erfolgreich gespeichert',
           'Es ist ein Fehler aufgetreten. Änderungen wurden nicht gespeichert',
-          () => this.leaveEditMode()
+          () => this.leaveEditMode(),
+        () => {},
+        () => this.store.dispatch(EditSkillActions.reset())
         )
       ));
   }
@@ -109,7 +167,12 @@ export class KexProfileSkillComponent implements OnInit, OnDestroy {
       this.$deleteSkillLoadState.pipe(
       ).subscribe(state => this.coreService.handleRequestState(state,
         'Fähigkeit wurde gelöscht',
-        'Es ist ein Fehler aufgetreten. Fähigkeit wurde nicht gelöscht.')
+        'Es ist ein Fehler aufgetreten. Fähigkeit wurde nicht gelöscht.',
+        () => this.profileService.loadSkills(),
+        () => {},
+        () => this.store.dispatch(DeleteSkillActions.reset())
+        )
+
       ));
   }
 
@@ -122,7 +185,9 @@ export class KexProfileSkillComponent implements OnInit, OnDestroy {
           () => {
             this.leaveEditMode();
             this.profileService.loadSkills();
-          }
+          },
+        () => {},
+        () => this.store.dispatch(AddSkillActions.reset())
         )
       ));
   }
@@ -132,7 +197,11 @@ export class KexProfileSkillComponent implements OnInit, OnDestroy {
       this.profileService.$updateVisibilitySkillLoadState.pipe(
       ).subscribe(state => this.coreService.handleRequestState(state,
         '',
-        'Es ist ein Fehler aufgetreten. Sichtbarkeit wurde nicht aktualisiert')
+        'Es ist ein Fehler aufgetreten. Sichtbarkeit wurde nicht aktualisiert',
+        () => {},
+        () => {},
+        () => this.store.dispatch(UpdateVisibilitySkillActions.reset())
+        )
       ));
   }
